@@ -10,8 +10,8 @@
     "lodash": "^4.17.21",
     "lodash.template": "^4.5.0",
     "mixin-deep": "^2.0.1",
-    "react": "^16.13.1",
-    "react-dom": "^16.13.1",
+    "react": "^17.0.0",
+    "react-dom": "^17.0.0",
     "react-router-dom": "^5.2.0",
     "react-scripts": "^3.4.1",
     "set-value": "^4.0.1"
