@@ -1,87 +1,72 @@
-import React, { Fragment } from 'react';
+import React, { useState } from 'react';
 import Paper from '@material-ui/core/Paper';
 import CapitalizeFirstLetter from '../UtilComponents/CapitalizeFirstLetter.js';
 
 const API_KEY = `${process.env.REACT_APP_API_KEY}`;
-const Location =  "https://extreme-ip-lookup.com/json/";
 
-class LocalWeatherContainer extends React.Component {
-  state = {
-    city: undefined,
-    temperature: undefined,
-    icon: undefined,
-    description: '',
-    humidity: undefined,
-    windSpeed: undefined,
-    cloud: undefined,
-  };
+function LocalWeatherContainer() {
+  const [location, setLocation] = useState(null);
+  const [apiResponse, setApiResonse] = useState(null);
 
-  async componentDidMount() {
-    // Get Location
-    const location_call = await fetch(Location);
-    const location_data = await location_call.json();
-    const city = await location_data.city;
-    // Get Weather with location data
-    const dataRes = await fetch(`https://api.openweathermap.org/data/2.5/weather?q=${city}&appid=${API_KEY}&units=imperial`);
-    const weather_data = await dataRes.json();
-    if (weather_data.name) {
-      this.setState({
-        city: weather_data.name,
-        temperature: Math.round(weather_data.main.temp),
-        icon: weather_data.weather[0].icon,
-        description: weather_data.weather[0].description,
-        humidity: weather_data.main.humidity,
-        windSpeed: Math.round(weather_data.wind.speed),
-        cloud: weather_data.clouds.all,
-      });
+  function handleGetLocation() {
+    if (navigator.geolocation) {
+      navigator.geolocation.getCurrentPosition(success, error);
     } else {
-      this.setState({
-        city: undefined,
-        temperature: undefined,
-        icon: undefined,
-        description: undefined,
-        humidity: undefined,
-        windSpeed: undefined,
-        cloud: undefined,
-      });
-    };
-  };
-
-  render () {
-    return (
-      <Fragment>
-        {this.state.city &&
-          <Paper className="local-weather-ticker">
-            <span className="local-weather-ticker-header">{this.state.city} local weather</span>
-            <div>
-              <span className="local-weather-category">Conditions</span>
-              {this.state.temperature}&deg;F
-                { this.state.icon &&
-                  <img className="weather-icon" src={"https://openweathermap.org/img/w/" + this.state.icon + ".png"} alt="Weather icon">
-                  </img>
-                }
-                {
-                  this.state.description &&
-                  <CapitalizeFirstLetter text={this.state.description} />
-                }
-            </div>
-            <div>
-              <span className="local-weather-category">Humidity</span>
-              {this.state.humidity}%
-            </div>
-            <div>
-              <span className="local-weather-category">Wind</span>
-              {this.state.windSpeed}mph
-            </div>
-            <div>
-              <span className="local-weather-category">Cloud</span>
-              {this.state.cloud}%
-            </div>
-          </Paper>
-        }
-      </Fragment>
-    )
+      console.log("Geolocation not supported");
+    }
   }
+  function success(position) {
+    const latitude = position.coords.latitude;
+    const longitude = position.coords.longitude;
+    setLocation({ latitude, longitude });
+    console.log(`Latitude: ${latitude}, Longitude: ${longitude}`);
+    // Make API call to OpenWeatherMap
+    fetch(`https://api.openweathermap.org/data/2.5/weather?lat=${latitude}&lon=${longitude}&appid=${API_KEY}&units=imperial`)
+      .then(response => response.json())
+      .then(data => {
+        setApiResonse(data);
+        console.log(data);
+      })
+    .catch(error => console.log(error));
+  }
+  function error() {
+    console.log("Unable to retrieve your location");
+  }
+  handleGetLocation();
+
+  return (
+    <div>
+      {apiResponse ? (
+        <Paper className="local-weather-ticker">
+          <span className="local-weather-ticker-header">{apiResponse.name} local weather</span>
+          <div>
+            <span className="local-weather-category">Conditions</span>
+            {Math.round(apiResponse.main.temp)}&deg;F
+              { apiResponse.weather[0].icon &&
+                <img className="weather-icon" src={"https://openweathermap.org/img/w/" + apiResponse.weather[0].icon + ".png"} alt="Weather icon">
+                </img>
+              }
+              {
+                apiResponse.weather[0].description &&
+                <CapitalizeFirstLetter text={apiResponse.weather[0].description} />
+              }
+          </div>
+          <div>
+            <span className="local-weather-category">Humidity</span>
+            {apiResponse.main.humidity}%
+          </div>
+          <div>
+            <span className="local-weather-category">Wind</span>
+            {Math.round(apiResponse.wind.speed)}mph
+          </div>
+          <div>
+            <span className="local-weather-category">Cloud</span>
+            {apiResponse.clouds.all}%
+          </div>
+        </Paper>
+      ) : null}
+    </div>
+  );
 }
 
 export default LocalWeatherContainer;
